+++ packages/analytics/analytics-config-store/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.16](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-config-store@1.2.15...@kong-ui-public/analytics-config-store@1.2.16) (2026-01-13)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-config-store
+
+
+
+
+
+## [1.2.15](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-config-store@1.2.14...@kong-ui-public/analytics-config-store@1.2.15) (2026-01-13)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-config-store
+
+
+
+
+
 ## [1.2.14](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-config-store@1.2.13...@kong-ui-public/analytics-config-store@1.2.14) (2026-01-09)
 
 **Note:** Version bump only for package @kong-ui-public/analytics-config-store