+++ packages/analytics/analytics-metric-provider/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [11.1.13](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@11.1.12...@kong-ui-public/analytics-metric-provider@11.1.13) (2026-01-13)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
+## [11.1.12](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@11.1.11...@kong-ui-public/analytics-metric-provider@11.1.12) (2026-01-13)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
 ## [11.1.11](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@11.1.10...@kong-ui-public/analytics-metric-provider@11.1.11) (2026-01-09)
 
 **Note:** Version bump only for package @kong-ui-public/analytics-metric-provider