+++ packages/analytics/analytics-utilities/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [12.8.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-utilities@12.7.3...@kong-ui-public/analytics-utilities@12.8.0) (2026-01-13)
+
+
+### Features
+
+* mcp info types on explore ([#2822](https://github.com/Kong/public-ui-components/issues/2822)) ([76d9840](https://github.com/Kong/public-ui-components/commit/76d9840dc61638157987afc56eaf57a79f1aa836))
+
+
+
+
+
+## [12.7.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-utilities@12.7.2...@kong-ui-public/analytics-utilities@12.7.3) (2026-01-13)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-utilities
+
+
+
+
+
 ## [12.7.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-utilities@12.7.1...@kong-ui-public/analytics-utilities@12.7.2) (2026-01-09)