+++ packages/analytics/dashboard-renderer/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [8.5.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/dashboard-renderer@8.4.21...@kong-ui-public/dashboard-renderer@8.5.0) (2026-01-13)
+
+
+### Features
+
+* mcp info types on explore ([#2822](https://github.com/Kong/public-ui-components/issues/2822)) ([76d9840](https://github.com/Kong/public-ui-components/commit/76d9840dc61638157987afc56eaf57a79f1aa836))
+
+
+
+
+
+## [8.4.21](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/dashboard-renderer@8.4.20...@kong-ui-public/dashboard-renderer@8.4.21) (2026-01-13)
+
+**Note:** Version bump only for package @kong-ui-public/dashboard-renderer
+
+
+
+
+
 ## [8.4.20](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/dashboard-renderer@8.4.19...@kong-ui-public/dashboard-renderer@8.4.20) (2026-01-09)
 
 **Note:** Version bump only for package @kong-ui-public/dashboard-renderer